Taller_s1/AF_MachineLearning/requirements.txt
azure-functions
pandas
pyodbc
sklearn
